# swagger.yaml
openapi: 3.0.0
info:
  title: Aeropaye API
  contact: {}
  version: '1.0'
servers:
- url: https://aeropaye-api.herokuapp.com/api/v1
  variables: {}
paths:
...
  /flights:
    get:
      tags:
      - Misc
      summary: Get All Flights
      operationId: GetAllFlights
      parameters:
      - name: api-key
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: $2a$10$.1Efok0/oDbPx6ZYUkUaBOWa0wzyXEIAmSweVlRl/CX6aM7sAf6pq
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
security:
- {}
...